moving check if claimed flow to api
